+++ IronSource/src/main/java/com/applovin/mediation/adapters/IronSourceMediationAdapter.java
@@ -33,17 +33,23 @@ import com.ironsource.mediationsdk.utils.IronSourceConstants;
 import com.ironsource.mediationsdk.utils.IronSourceUtils;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
+
+import androidx.annotation.Nullable;
 
 public class IronSourceMediationAdapter
         extends MediationAdapterBase
         implements MaxSignalProvider, MaxInterstitialAdapter, MaxRewardedAdapter, MaxAdViewAdapter
 {
-    private static final IronSourceRouter ROUTER      = new IronSourceRouter();
-    private static final AtomicBoolean    INITIALIZED = new AtomicBoolean();
+    private static final IronSourceRouter ROUTER                           = new IronSourceRouter();
+    private static final AtomicBoolean    INITIALIZED                      = new AtomicBoolean();
+    private static final List<String>     loadedAdViewPlacementIdentifiers = Collections.synchronizedList( new ArrayList<>() );
 
     private String                   mRouterPlacementIdentifier;
+    @Nullable
+    private String                   adViewPlacementIdentifier;
     private ISDemandOnlyBannerLayout adView;
 
     public IronSourceMediationAdapter(final AppLovinSdk sdk) { super( sdk ); }
@@ -105,6 +111,14 @@ public class IronSourceMediationAdapter
     @Override
     public void onDestroy()
     {
+        if ( adViewPlacementIdentifier != null )
+        {
+            log( "Destroying adview with instance ID: " + adViewPlacementIdentifier );
+
+            IronSource.destroyISDemandOnlyBanner( adViewPlacementIdentifier );
+            loadedAdViewPlacementIdentifiers.remove( adViewPlacementIdentifier );
+        }
+
         ROUTER.removeAdapter( this, mRouterPlacementIdentifier );
     }
 
@@ -234,20 +248,30 @@ public class IronSourceMediationAdapter
 
         final String bidResponse = parameters.getBidResponse();
         final boolean isBiddingAd = AppLovinSdkUtils.isValidString( bidResponse );
-        final String instanceId = parameters.getThirdPartyAdPlacementId();
 
-        log( "Loading " + ( isBiddingAd ? "bidding " : "" ) + adFormat.getLabel() + " ad for instance ID: " + instanceId );
+        log( "Loading " + ( isBiddingAd ? "bidding " : "" ) + adFormat.getLabel() + " ad for instance ID: " + parameters.getThirdPartyAdPlacementId() );
+
+        // IronSource does not support b2b with same instance id for banners/MRECs
+        if ( loadedAdViewPlacementIdentifiers.contains( parameters.getThirdPartyAdPlacementId() ) )
+        {
+            log( "AdView ad failed to load for instance ID: " + parameters.getThirdPartyAdPlacementId() + ". An ad with the same instance ID is already loaded" );
+            listener.onAdViewAdLoadFailed( new MaxAdapterError( MaxAdapterError.INTERNAL_ERROR.getCode(), MaxAdapterError.INTERNAL_ERROR.getMessage(), 0, "An ad with the same instance ID is already loaded" ) );
+
+            return;
+        }
+
+        adViewPlacementIdentifier = parameters.getThirdPartyAdPlacementId(); // Set it only if it is not an instance id of an already loaded ad to avoid destroying the currently showing ad
 
         adView = IronSource.createBannerForDemandOnly( activity, toISBannerSize( adFormat ) );
         adView.setBannerDemandOnlyListener( new AdViewListener( listener ) );
 
         if ( isBiddingAd )
         {
-            IronSource.loadISDemandOnlyBannerWithAdm( activity, adView, instanceId, bidResponse );
+            IronSource.loadISDemandOnlyBannerWithAdm( activity, adView, adViewPlacementIdentifier, bidResponse );
         }
         else
         {
-            IronSource.loadISDemandOnlyBanner( activity, adView, instanceId );
+            IronSource.loadISDemandOnlyBanner( activity, adView, adViewPlacementIdentifier );
         }
     }
 
@@ -580,6 +604,8 @@ public class IronSourceMediationAdapter
         @Override
         public void onBannerAdShown(String instanceId)
         {
+            loadedAdViewPlacementIdentifiers.add( instanceId );
+
             log( "AdView ad displayed for instance ID: " + instanceId );
             listener.onAdViewAdDisplayed();
         }